read debug setting from envs
