+++ README.md
@@ -27,3 +27,16 @@ make
 ```
 make build
 ```
+
+## Настройки
+
+При запуске бот считывает настройки из переменных окружения и из файла .env в рабочей директории.
+Если настройка присутствует и в переменной окружения и в файле, то приоритет отдается переменной окружения.
+
+### Обязательные настройки
+
+- TOKEN -- строка с токеном для авторизации бота в Телеграме, получается при создании бота от [BotFather](https://telegram.me/BotFather)
+
+### Опциональные настройки
+
+- DEBUG -- присвойте значение "on" (регистр не важен), чтобы включить отладочные сообщения при запросах к API Телеграма. Любые другие значения, в том числе если настройка отсутствует, воспринимаются как отключено
